move devices to utils, change paths
