+++ src/components/Landing/Main/Main-styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import devices from '../../devices';
+import devices from '../../../utils/devices';
 
 export const MainContainer = styled.div`
   max-width: 100%;
