+++ src/components/Onboarding/StudentForm.js
@@ -19,7 +19,7 @@ import {
   Logo,
 } from '../Landing/Landing-styles';
 import { chooseUserRole } from '../../state/actions/authenticationActions';
-import { countries } from '../Forms/countries';
+import { countries } from '../../utils/countries';
 
 const NavLogo = styled(Logo)`
   a {
