+++ src/utils/devices.js
